requirements.txt
fastapi
uvicorn[standard]
motor                  # async MongoDB driver
python-dotenv
pydantic
python-jose[cryptography]   # JWT handling
passlib[bcrypt]             # password hashing